Set manual doc deployment
The docs should be updated when a new release is made. There is no point
to update the docs every time when some new changes are merged into the
master branch. The proposed change enables manual run of the
deployment, that should be run when a new release is issued.